add testing packages
